@@ -18,7 +18,7 @@
         "facebook"
     ],
     "require": {
-        "php": "^5.6 || ^7.0 || ^8.0",
+        "php": ">=7.3",
         "league/oauth2-client": "^2.0"
     },
     "require-dev": {
