@@ -4,19 +4,19 @@ namespace League\OAuth2\Client\Grant;
 
 class FbExchangeToken extends AbstractGrant
 {
-    public function __toString()
+    public function __toString(): string
     {
         return 'fb_exchange_token';
     }
 
-    protected function getRequiredRequestParameters()
+    protected function getRequiredRequestParameters(): array
     {
         return [
             'fb_exchange_token',
         ];
     }
 
-    protected function getName()
+    protected function getName(): string
     {
         return 'fb_exchange_token';
     }
