@@ -5,6 +5,7 @@ namespace League\OAuth2\Client\Provider;
 use League\OAuth2\Client\Token\AccessToken;
 use League\OAuth2\Client\Provider\Exception\FacebookProviderException;
 use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
+use League\OAuth2\Client\Token\AccessTokenInterface;
 use Psr\Http\Message\ResponseInterface;
 
 /**
@@ -17,36 +18,36 @@ class Facebook extends AbstractProvider
      *
      * @const string
      */
-    const BASE_FACEBOOK_URL = 'https://www.facebook.com/';
+    protected const BASE_FACEBOOK_URL = 'https://www.facebook.com/';
 
     /**
      * Beta tier URL of the Graph API.
      *
      * @const string
      */
-    const BASE_FACEBOOK_URL_BETA = 'https://www.beta.facebook.com/';
+    protected const BASE_FACEBOOK_URL_BETA = 'https://www.beta.facebook.com/';
 
     /**
      * Production Graph API URL.
      *
      * @const string
      */
-    const BASE_GRAPH_URL = 'https://graph.facebook.com/';
+    protected const BASE_GRAPH_URL = 'https://graph.facebook.com/';
 
     /**
      * Beta tier URL of the Graph API.
      *
      * @const string
      */
-    const BASE_GRAPH_URL_BETA = 'https://graph.beta.facebook.com/';
+    protected const BASE_GRAPH_URL_BETA = 'https://graph.beta.facebook.com/';
 
     /**
      * Regular expression used to check for graph API version format
      *
      * @const string
      */
-    const GRAPH_API_VERSION_REGEX = '~^v\d+\.\d+$~';
-    
+    protected const GRAPH_API_VERSION_REGEX = '~^v\d+\.\d+$~';
+
     /**
      * The Graph API version to use for requests.
      *
@@ -74,7 +75,9 @@ class Facebook extends AbstractProvider
         if (empty($options['graphApiVersion'])) {
             $message = 'The "graphApiVersion" option not set. Please set a default Graph API version.';
             throw new \InvalidArgumentException($message);
-        } elseif (!preg_match(self::GRAPH_API_VERSION_REGEX, $options['graphApiVersion'])) {
+        }
+
+        if (!preg_match(self::GRAPH_API_VERSION_REGEX, $options['graphApiVersion'])) {
             $message = 'The "graphApiVersion" must start with letter "v" followed by version number, ie: "v2.4".';
             throw new \InvalidArgumentException($message);
         }
@@ -86,22 +89,22 @@ class Facebook extends AbstractProvider
         }
     }
 
-    public function getBaseAuthorizationUrl()
+    public function getBaseAuthorizationUrl(): string
     {
-        return $this->getBaseFacebookUrl().$this->graphApiVersion.'/dialog/oauth';
+        return $this->getBaseFacebookUrl() . $this->graphApiVersion . '/dialog/oauth';
     }
 
-    public function getBaseAccessTokenUrl(array $params)
+    public function getBaseAccessTokenUrl(array $params): string
     {
-        return $this->getBaseGraphUrl().$this->graphApiVersion.'/oauth/access_token';
+        return $this->getBaseGraphUrl() . $this->graphApiVersion . '/oauth/access_token';
     }
 
-    public function getDefaultScopes()
+    public function getDefaultScopes(): array
     {
         return ['public_profile', 'email'];
     }
 
-    public function getResourceOwnerDetailsUrl(AccessToken $token)
+    public function getResourceOwnerDetailsUrl(AccessToken $token): string
     {
         $fields = [
             'id', 'name', 'first_name', 'last_name',
@@ -116,11 +119,13 @@ class Facebook extends AbstractProvider
 
         $appSecretProof = AppSecretProof::create($this->clientSecret, $token->getToken());
 
-        return $this->getBaseGraphUrl().$this->graphApiVersion.'/me?fields='.implode(',', $fields)
-                        .'&access_token='.$token.'&appsecret_proof='.$appSecretProof;
+        return $this->getBaseGraphUrl()
+            . $this->graphApiVersion
+            . '/me?fields=' . implode(',', $fields)
+            . '&access_token=' . $token . '&appsecret_proof=' . $appSecretProof;
     }
 
-    public function getAccessToken($grant = 'authorization_code', array $params = [])
+    public function getAccessToken($grant = 'authorization_code', array $params = []): AccessTokenInterface
     {
         if (isset($params['refresh_token'])) {
             throw new FacebookProviderException('Facebook does not support token refreshing.');
@@ -131,39 +136,35 @@ class Facebook extends AbstractProvider
 
     /**
      * Exchanges a short-lived access token with a long-lived access-token.
-     *
-     * @param string $accessToken
-     *
-     * @return \League\OAuth2\Client\Token\AccessToken
-     *
-     * @throws FacebookProviderException
      */
-    public function getLongLivedAccessToken($accessToken)
+    public function getLongLivedAccessToken(string $accessToken): AccessTokenInterface
     {
         $params = [
-            'fb_exchange_token' => (string) $accessToken,
+            'fb_exchange_token' => $accessToken,
         ];
 
         return $this->getAccessToken('fb_exchange_token', $params);
     }
 
-    protected function createResourceOwner(array $response, AccessToken $token)
+    protected function createResourceOwner(array $response, AccessToken $token): FacebookUser
     {
         return new FacebookUser($response);
     }
 
-    protected function checkResponse(ResponseInterface $response, $data)
+    protected function checkResponse(ResponseInterface $response, $data): void
     {
-        if (!empty($data['error'])) {
-            $message = $data['error']['type'].': '.$data['error']['message'];
-            throw new IdentityProviderException($message, $data['error']['code'], $data);
+        if (empty($data['error'])) {
+            return;
         }
+
+        $message = $data['error']['type'] . ': ' . $data['error']['message'];
+        throw new IdentityProviderException($message, $data['error']['code'], $data);
     }
 
     /**
      * @inheritdoc
      */
-    protected function getContentType(ResponseInterface $response)
+    protected function getContentType(ResponseInterface $response): string
     {
         $type = parent::getContentType($response);
 
@@ -182,20 +183,16 @@ class Facebook extends AbstractProvider
 
     /**
      * Get the base Facebook URL.
-     *
-     * @return string
      */
-    private function getBaseFacebookUrl()
+    private function getBaseFacebookUrl(): string
     {
         return $this->enableBetaMode ? static::BASE_FACEBOOK_URL_BETA : static::BASE_FACEBOOK_URL;
     }
 
     /**
      * Get the base Graph API URL.
-     *
-     * @return string
      */
-    private function getBaseGraphUrl()
+    private function getBaseGraphUrl(): string
     {
         return $this->enableBetaMode ? static::BASE_GRAPH_URL_BETA : static::BASE_GRAPH_URL;
     }
